BETA: New implementation of geonengine using an abstraction layer allowing support for different engines. 	* Patch by: FBIagent 	* Tested by: lion, oscard, Konstantinos, Szponiasty, dEvilKinG, skyshadow, janiko, pandragon, badboy29, midnex, valanths1990 	* Reviewed by: Zoey76, lion, BiggBoss

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/pathfinding/AbstractNodeLoc.java
@@ -27,7 +27,7 @@ public abstract class AbstractNodeLoc
 	
 	public abstract int getY();
 	
-	public abstract short getZ();
+	public abstract int getZ();
 	
 	public abstract void setZ(short z);
