+++ L2J_Server_BETA/java/com/l2jserver/gameserver/pathfinding/geonodes/GeoNodeLoc.java
@@ -50,7 +50,7 @@ public class GeoNodeLoc extends AbstractNodeLoc
 	}
 	
 	@Override
-	public short getZ()
+	public int getZ()
 	{
 		return _z;
 	}
